services/srv-agendamentos/requirements.txt
fastapi
uvicorn[standard]
sqlalchemy
psycopg2-binary
pydantic-settings
pika

python-jose[cryptography]
passlib[bcrypt]